Remove four unused modules, extract chat service
Removes Clinical Trial Matcher, Outlive Checklist, Undiagnosed Diseases
Network, and CBT Recovery Coach modules (8 files, ~1944 lines) along
with their navigation entries in HomeView and ContentView.

The Chat with My Records feature previously depended on chat-only
methods that lived inside the misleadingly named ClinicalTrialMatcher
class. Those methods (convertRecordsToMedicalText, createChatSession,
generateChatURL, formatting helpers) and the MedicalTextView preview
are extracted into a new ChatRecordsService so the kept module keeps
working.

Closes #13.
